+++ src/HetrickCV.hpp
@@ -23,7 +23,9 @@ struct FlipFlopWidget : ModuleWidget { FlipFlopWidget();};
 struct Boolean3Widget : ModuleWidget { Boolean3Widget();};
 struct GateJunctionWidget : ModuleWidget { GateJunctionWidget();};
 struct LogicCombineWidget : ModuleWidget { LogicCombineWidget();};
+
 struct ComparatorWidget : ModuleWidget { ComparatorWidget();};
+struct DeltaWidget : ModuleWidget { DeltaWidget();};
 
 struct AnalogToDigitalWidget : ModuleWidget { AnalogToDigitalWidget();};
 struct DigitalToAnalogWidget : ModuleWidget { DigitalToAnalogWidget();};